+++ src/ui/components/usersListTable/UsersListTable.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 
-import { getImageUrl } from "../../../utils";
-import { User, UserRecipeType } from "../../pages/usersList/types"
-import { formatLongDate } from "../../utils/date"
+import { formatLongDate, getImageUrl } from '../../../utils'
+import { User, UserRecipeType } from '../../pages/usersList/types'
 import NoUsers from "../noUsers/NoUsers"
 import PhoneDisplay from "../phoneNumber/PhoneNumber"
 import "./UsersListTable.scss"
